Rename PORT-TEMPLATE to ISSUE_TEMPLATE/device-port.md (#53)
This will make this a valid, possible issue template as GitHub suggests it. You
can use it by adding `?template=device-port.md` to the new issue URL, so the
resulting URL would be:
https://github.com/Halium/projectmanagement/issues/new?template=device-port.md

Also see https://github.com/blog/2495-multiple-issue-and-pull-request-templates
for more information. It's a quite new feature from 25th January, 2018.

Thanks to @JBBgameich for pointing out that this feature exists.